Move GLB computation into `higher_ranked`
